+++ changelog.md
@@ -5,6 +5,7 @@ Saves from 2.5 are not compatible with 2.6.
 ## Features/Improvements
 
 * **[Campaign]** Ground units can now be transferred by road. See https://github.com/Khopa/dcs_liberation/wiki/Unit-Transfers for more information.
+* **[Campaign]** Ground units can no longer be sold. To move units to a new location, transfer them.
 
 ## Fixes